requirements.txt
certifi==2022.9.24
python-dotenv==0.20.0
requests==2.27.1
terminaltables==3.1.10
